lib: rename BufferList.js to buffer_list.js
