+++ node.gyp
@@ -147,7 +147,7 @@
       'lib/internal/stream_base_commons.js',
       'lib/internal/streams/lazy_transform.js',
       'lib/internal/streams/async_iterator.js',
-      'lib/internal/streams/BufferList.js',
+      'lib/internal/streams/buffer_list.js',
       'lib/internal/streams/duplex_base.js',
       'lib/internal/streams/duplexpair.js',
       'lib/internal/streams/legacy.js',
